Refactor: Reorganize packages and modules under `io.github.kdroidfilter.androidappstorekit`. Update imports, namespaces, and build scripts to align with new structure.
