+++ aptoide/api/src/commonMain/kotlin/io/github/kdroidfilter/androidappstorekit/aptoide/api/constants/ApiConstants.kt
@@ -1,0 +1,4 @@
+package io.github.kdroidfilter.androidappstorekit.aptoide.api.constants
+
+internal const val BASE_APTOIDE_API_URL = "https://ws2.aptoide.com/api/7"
+internal const val APP_GET_META_PATH = "/app/getMeta"
