--- aptoide/src/commonMain/kotlin/com/kdroid/aptoide/constants/ApiConstants.kt
@@ -1,4 +1,0 @@
-package com.kdroid.aptoide.constants
-
-internal const val BASE_APTOIDE_API_URL = "https://ws2.aptoide.com/api/7"
-internal const val APP_GET_META_PATH = "/app/getMeta"
