--- core/build.gradle.kts
@@ -1,120 +1,0 @@
-import com.vanniktech.maven.publish.SonatypeHost
-
-plugins {
-    alias(libs.plugins.multiplatform)
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.kotlinx.serialization)
-    alias(libs.plugins.vannitktech.maven.publish)
-}
-
-group = "com.krdoid.gplayscrapper.core"
-val ref = System.getenv("GITHUB_REF") ?: ""
-val version = if (ref.startsWith("refs/tags/")) {
-    val tag = ref.removePrefix("refs/tags/")
-    if (tag.startsWith("v")) tag.substring(1) else tag
-} else "dev"
-
-kotlin {
-    jvmToolchain(11)
-    androidTarget {
-        publishLibraryVariants("release")
-    }
-    wasmJs { browser() }
-
-    jvm()
-
-    linuxX64 {
-        binaries.staticLib {
-            baseName = "shared"
-        }
-    }
-
-    mingwX64 {
-        binaries.staticLib {
-            baseName = "shared"
-        }
-    }
-
-    listOf(
-        iosX64(),
-        iosArm64(),
-        iosSimulatorArm64()
-    ).forEach {
-        it.binaries.framework {
-            baseName = "core"
-            isStatic = true
-        }
-    }
-
-    listOf(
-        macosX64(),
-        macosArm64()
-    ).forEach {
-        it.binaries.framework {
-            baseName = "core"
-            isStatic = true
-        }
-    }
-
-    sourceSets {
-        commonMain.dependencies {
-            implementation(libs.kotlinx.serialization.json)
-        }
-
-        commonTest.dependencies {
-            implementation(kotlin("test"))
-        }
-    }
-
-    //https://kotlinlang.org/docs/native-objc-interop.html#export-of-kdoc-comments-to-generated-objective-c-headers
-    targets.withType<org.jetbrains.kotlin.gradle.plugin.mpp.KotlinNativeTarget> {
-        compilations["main"].compilerOptions.options.freeCompilerArgs.add("-Xexport-kdoc")
-    }
-}
-
-android {
-    namespace = "com.krdoid.gplayscrapper.core"
-    compileSdk = 35
-
-    defaultConfig {
-        minSdk = 21
-    }
-}
-
-mavenPublishing {
-    coordinates(
-        groupId = "io.github.kdroidfilter",
-        artifactId = "gplayscrapper-core",
-        version = version.toString()
-    )
-
-    pom {
-        name.set("GPlay Scrapper Core Library")
-        description.set("Core module for GPlay Scrapper Library containing model classes")
-        inceptionYear.set("2024")
-        url.set("https://github.com/kdroidFilter/GPlay-Scrapper")
-
-        licenses {
-            license {
-                name.set("MIT")
-                url.set("https://opensource.org/licenses/MIT")
-            }
-        }
-
-        developers {
-            developer {
-                id.set("kdroidFilter")
-                name.set("Elie Gambache")
-                email.set("elyahou.hadass@gmail.com")
-            }
-        }
-
-        scm {
-            url.set("https://github.com/kdroidFilter/GPlay-Scrapper")
-        }
-    }
-
-    publishToMavenCentral(SonatypeHost.CENTRAL_PORTAL)
-
-    signAllPublications()
-}
