+++ gplay/scrapper/src/commonMain/kotlin/io/github/kdroidfilter/androidappstorekit/gplay/scrapper/constants/ApiConstants.kt
@@ -1,0 +1,4 @@
+package io.github.kdroidfilter.androidappstorekit.gplay.scrapper.constants
+
+internal const val BASE_PLAY_STORE_URL = "https://play.google.com"
+internal const val DETAIL_PATH = "/store/apps/details"
