+++ gplay/scrapper/src/commonMain/kotlin/io/github/kdroidfilter/androidappstorekit/gplay/scrapper/constants/RegexConstants.kt
@@ -1,0 +1,5 @@
+package io.github.kdroidfilter.androidappstorekit.gplay.scrapper.constants
+
+// Regex for datasets
+internal val keyRegex = Regex("(ds:\\d+)")
+internal val valueRegex = Regex("data:([\\s\\S]*?),\\s*sideChannel:")
