--- gplayscrapper/src/commonMain/kotlin/com/kdroid/gplayscrapper/constants/ApiConstants.kt
@@ -1,4 +1,0 @@
-package com.kdroid.gplayscrapper.constants
-
-internal const val BASE_PLAY_STORE_URL = "https://play.google.com"
-internal const val DETAIL_PATH = "/store/apps/details"
