--- gplayscrapper/src/commonMain/kotlin/com/kdroid/gplayscrapper/constants/RegexConstants.kt
@@ -1,5 +1,0 @@
-package com.kdroid.gplayscrapper.constants
-
-// Regex for datasets
-internal val keyRegex = Regex("(ds:\\d+)")
-internal val valueRegex = Regex("data:([\\s\\S]*?),\\s*sideChannel:")
